Add logic to use a specified entity attribute ID
The code previously relied on the first unique tracked entity attribute
ID to use for checking for duplicates and searching for mother's. With
the addition of the system ID which uniquely identifies mothers within
the system, this will now be the prefered unique ID as it is far more
reliable compared to the mother's phone number.

However for backward compatibility if a preferred unique ID is not
specified this will default to using the first unique ID processed as
was the functionality before.

MOM-455 MOM-460