Rename variable
To match its new context with the following changes.
The changes to follow will add a new identifier. This unique entity
attribute ID maybe specified in the options coming from the openhim
mediator to be used for searches, etc.

Therefore the uniqueTrackedEntityAttributeID will no longer just be the
first unique identifier it finds.

MOM-455 MOM-460