Fixed non-unique import summary conflict check
With the addition of a new unique identifier it is now possible for
multiple valid conflicts to occur in the import summary. Previously
there should only have been one valid conflict (Mother's cell phone
number).

The conflicts occur because these are unique identifiers and will
therefore be passed through in every message related to a specific
mother (theoretically).

This check ensure that the conflicts in the import summary are only due
to these unique identifiers and not due to an error.
Conflicts due the unique attributes are ignored and the tracked entity
instance can be updated.

MOM-455 MOM-460