Add eventID lookup for duplicate prevention
The eventID is a data element added to the Event resource in the
momconnect DHIS2 instance. A UUID is generated and given to each message
from Praekelt which will be the data element value. When new messages
are received this allows us to now lookup existing events and make sure
that the UUID doesn't already exist, thereby preventing duplicate
messages being processed.

This change is backwards compatible. If the uniqueDataElementID option
is specified in the Populator instantiationt then this accurate method
of de-duplication will be used. However, if the it is left out the
previous method using the duplicationThreshold value will be used. This
method is not accurate. The eventID method is therefore the preferred
method.

MOM-464