Rename type cache and populator option variable
The uniqueDataElement is the UID (name) of a field on DHIS that is used
here to uniquely Identify events.
This variable differed from the Type-cache and Populator option. This
was confusing.
The names were different to try differentiate between the field coming
in from the Populator options and the type-Cache during development to
prevent accidental usage of the Option variable which may have resulted
in inconsistentcies as these could theoretically be different.
But that shouldn't be an issue now.

MOM-464